+++ R/ctrLoadQueryIntoDb.R
@@ -6,10 +6,11 @@
 #' and stores it in a collection in a database. Main function
 #' of \link{ctrdata} for accessing registers.
 #' A collection can store trial information from different queries
-#' or different registers. Query details are stored in the
+#' or different registers.
+#' When at least one trial is loaded, query details are stored in the
 #' collection and can be accessed using \link{dbQueryHistory}.
-#' A previous query can be re-run, which replaces or adds trial
-#' records while keeping any user annotations of trial records.
+#' A previous query can be re-run, which replaces existing or adds new trial
+#' records while keeping any user annotations of existing trial records.
 #'
 #' @param queryterm Either a single string with the full URL of a search
 #' query in a register, e.g. one of the strings returned by
@@ -36,7 +37,7 @@
 #' this query was run the last time.
 #' This parameter takes precedence over \code{queryterm}.
 #' For "EUCTR" and "CTIS", updates are available only for the last seven days;
-#' the query is run again if more time has passed since it was run last.
+#' the query is run again fully if more time has passed since it was run last.
 #'
 #' @param forcetoupdate If \code{TRUE}, run again the query
 #' given in \code{querytoupdate}, irrespective of when it was
@@ -74,11 +75,12 @@
 #' This somewhat time-consuming. Default is \code{FALSE}.
 #'
 #' @param euctrprotocolsall If \code{TRUE}, load all available records of
-#' protocol-related data (that is, versions from all EU Member States and any
-#' third country where the trial is conducted); if \code{FALSE}, only a single
-#' record per trial is loaded, to accelerate loading. Default is \code{TRUE},
-#' but only for backwards consistency; for new collections, \code{FALSE} is
-#' the recommended setting, unless there are questions about differences
+#' protocol-related data (that is, available versions from any EU Member States
+#' and any third country where the trial is or was conducted); if \code{FALSE},
+#' to accelerate loading, load only a single EU Member State record per trial
+#' (or no record, if records exist only for third countries).
+#' Default is \code{TRUE}, for backwards consistency; for new collections,
+#' \code{FALSE} is recommended, unless there are questions about differences
 #' between Member States' protocol versions of a trial such as dates or outcomes
 #' of an authorisation decision or an ethics opinion, global status and end.
 #'
@@ -98,7 +100,8 @@
 #' move the current CTIS record into an array \code{history} with the record
 #' which holds one or more historic versions, before updating the rest of the
 #' record from CTIS. Default is \code{FALSE}, because this is a time-consuming
-#' operation. See "Historic versions" in \code{vignette("ctrdata_summarise")}.
+#' operation. See "Historic versions..." in vignette
+#' \href{../doc/ctrdata_summarise.html}{ctrdata_summarise}.
 #'
 #' @param annotation.text Text to be including into the field
 #' "annotation" in the records retrieved with the query