+++ R/ctrLoadQueryIntoDbEuctr.R
@@ -161,7 +161,7 @@ ctrLoadQueryIntoDbEuctr <- function(
       dir.create(documents.path, recursive = TRUE, showWarnings = FALSE),
       silent = TRUE)
     if (!inherits(createdDir, "try-errror") && createdDir) {
-      message("Created directory ", documents.path)
+      message("- Created directory ", documents.path)
     } else {
       warning("Directory could not be created for 'documents.path' ",
               documents.path, ", ignored", call. = FALSE)
@@ -176,7 +176,7 @@ ctrLoadQueryIntoDbEuctr <- function(
   # inform user
   message(
     "- Downloading in ", resultsEuNumPages, " batch(es) (20 trials each; ",
-    "estimate: ", signif(resultsEuNumTrials * 0.12, 1L), " MB)")
+    "estimate: ", signif(resultsEuNumPages * 173 / 141, 1L), " s)...")
 
   # vector with URLs of all summaries
   urls <- sprintf(
@@ -195,11 +195,12 @@ ctrLoadQueryIntoDbEuctr <- function(
     ))
 
   # do download and saving
+  # print(system.time(
   resDf <- ctrMultiDownload(
     urls = urls,
     destfiles = fp,
-    verbose = verbose
-  )
+    verbose = verbose)
+  # )) # print system time
 
   if (nrow(resDf) != resultsEuNumPages) {
     message("Download from EUCTR failed; incorrect number of records")
@@ -235,10 +236,15 @@ ctrLoadQueryIntoDbEuctr <- function(
       if (!euctrprotocolsall) {
         p <- lapply(p, function(i) {
           if (length(i) == 1L) return(i)
+          # intentionally do not match 3RD
+          # (which would need to upper) as
+          # per euctrprotocolsall help text
           i <- i[i %in% countriesActive]
           w <- countriesPreferred %in% i
           w <- countriesPreferred[w][1]
-          if (is.na(w)) i[1] else w
+          if (!is.na(w)) return(w)
+          if (!is.na(i[1])) return(i[1])
+          return(NULL)
         })
       }
 
@@ -265,24 +271,36 @@ ctrLoadQueryIntoDbEuctr <- function(
     ))
 
   # inform user
-  message("- Downloading ", length(urls), " records of ", resultsEuNumTrials,
-          " trials (estimate: ", signif(length(urls) * 16 / 287, 1L), " s)")
+  message(
+    "- Downloading ", length(urls), " records of ", resultsEuNumTrials,
+    " trials (estimate: ", signif(length(urls) * 686 / 13608, 1L), " s)...")
+
+  # no trials for downloading
+  # (e.g., only GB and 3RD)
+  if (!length(urls)) {
+    # return
+    return(list(n = length(urls),
+                success = NULL,
+                failed = NULL))
+  }
 
   # do download and save
+  # print(system.time(
   resDf <- ctrMultiDownload(
     urls = urls,
     destfiles = fp,
-    verbose = verbose
-  )
+    verbose = verbose)
+  # )) # print system time
 
   ## convert euctr to ndjson -----------------------------------------------
 
   if (length(.ctrdataenv$ct) == 0L) initTranformers()
 
   # run conversion
-  message("- Converting to NDJSON (estimate: ",
-          signif(nrow(resDf) * 5.7 / 2100, 1L),
-          " s)... ", appendLF = FALSE)
+  message(
+    "- Converting to NDJSON (estimate: ",
+    signif(nrow(resDf) * 32 / 13608, 1L),
+    " s)... ", appendLF = FALSE)
 
   ti <- format(Sys.time(), "%Y-%m-%d %H:%M:%S")
   rl <- nrow(resDf)
@@ -516,8 +534,9 @@ ctrLoadQueryIntoDbEuctr <- function(
     on.exit(unlink(jsonFileList), add = TRUE)
 
     # iterate conversion over eudract result files
-    message("- Converting to NDJSON (estimate: ",
-            signif(length(xmlFileList) * 6.13 / 201, 1L), " s)...")
+    message(
+      "- Converting to NDJSON (estimate: ",
+      signif(length(xmlFileList) * 53 / 1889, 1L), " s)...")
 
     # run conversion
     # print(system.time(
@@ -567,7 +586,6 @@ ctrLoadQueryIntoDbEuctr <- function(
       )
 
     }) # sapply xmlFileList
-
     # )) # print system time
 
     # iterate over results files
@@ -606,7 +624,7 @@ ctrLoadQueryIntoDbEuctr <- function(
 
         warning(paste0(
           "Import of results failed for trial ", eudractNumber),
-        immediate. = TRUE)
+          immediate. = TRUE)
         res <- 0L
 
       } else {