+++ R/dbFindIdsUniqueTrials.R
@@ -13,6 +13,9 @@
 #' (and, for "EUCTR", between records for different Member States).
 #' Such differences are not considered by this function.
 #'
+#' Note that for "CTIS", the trial with the latest (highest) resubmission
+#' number (last two digits of `id`, clinical trial number) is identified.
+#'
 #' Note that the trial concept \link{f.isUniqueTrial} (which uses this
 #' function) can be calculated at the time of creating a data frame with
 #' \link{dbGetFieldsIntoDf}, which often may be the preferred approach.
@@ -116,7 +119,8 @@ dbFindIdsUniqueTrials <- function(
     # in the previously rbind'ed sets
     if (i > 1L && nrow(subDf)) {
 
-      # check for duplicates
+      # check for duplicates against
+      # identifiers in outSet so far
       dupes <- mapply(
         function(c1, c2) {
           tmpIs <- intersect(
@@ -136,18 +140,30 @@ dbFindIdsUniqueTrials <- function(
         },
         subDf[, colsToCheck, drop = FALSE],
         outSet[, colsToCheck, drop = FALSE],
-        SIMPLIFY = FALSE
+        SIMPLIFY = FALSE,
+        USE.NAMES = FALSE
       )
 
       # mangle dupes for marginal cases, e.g. one record
       dupes <- do.call(cbind, dupes)
       dupes <- as.data.frame(dupes)
 
-      # keep uniques
+      # keep uniques by removing rows from register set
       subDf <- subDf[rowSums(dupes) == 0L, , drop = FALSE]
-      rm(dupes)
 
     } # if
+
+    # special case: checking ctis resubmission number,
+    # keep trial record with latest = highest number
+    if (preferregister[i] == "CTIS") {
+
+      ctisBase <- sub("-[0-9]{2}$", "", subDf$CTIS)
+      ctisResub <- as.numeric(sub("^.+-([0-9]{2})$", "\\1", subDf$CTIS))
+      subDf <- subDf[order(ctisBase, -ctisResub), ]
+      ctisBase <- sub("-[0-9]{2}$", "", subDf$CTIS)
+      subDf <- subDf[!duplicated(ctisBase), , drop = FALSE]
+
+    }
 
     # add to output set
     outSet <- rbind(
@@ -562,6 +578,9 @@ dfFindUniqueEuctrRecord <- function(
     )
     include3rdcountrytrials <- TRUE
   }
+
+  # early exit
+  if (!any(df[["ctrname"]] == "EUCTR")) return(df)
 
   # count total
   totalEuctr <- unique(df[["a2_eudract_number"]])